Auth on test + cache + routes protection

+++ server/routers/recipeBook.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const database = require('../module/database');
 const conn = database.conn;
+const needAuth = require('../module/token').checkApiKey;
 var bodyParser = require('body-parser');
 
 router.use(bodyParser.json());
@@ -15,6 +16,18 @@ router.use(
 async function bookExist(bookId){
   let exists = await conn.query(`SELECT COUNT(id) FROM recipe_books WHERE id=?;`, [bookId]);
   return exists[0]['COUNT(id)'] <= 0;
+}
+
+async function userHasReadAccess(bookId, userId){
+  const isPublic = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND visibility = 1`, [bookId]);
+  const isOwner = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND owner = ? `, [bookId, userId]);
+  const isAuthorized = await conn.query(`SELECT id FROM recipe_books JOIN recipe_book_access ON recipe_books.id = recipe_book_access.bookId WHERE recipe_book_access.bookId = ? AND recipe_book_access.userId = ?;`, [bookId, userId]);
+  return isPublic.length >= 1 || isOwner.length >= 1 || isAuthorized.length >= 1;
+}
+
+async function isOwner(bookId, userId) {
+  const book = await conn.query(`SELECT id FROM recipe_books WHERE id = ? AND owner = ?`, [bookId, userId]);
+  return book.length == 1;
 }
 
 /**
@@ -34,7 +47,7 @@ async function bookExist(bookId){
  *   "400":
  *      description: "Name parameter is wrong"
  */
-router.post("/create", async (req, res) => {
+router.post("/create", needAuth, async (req, res) => {
     if (req.query.name === undefined){
         res.status(400).send("Need to specify a book name");
         return;
@@ -43,7 +56,7 @@ router.post("/create", async (req, res) => {
     const reg = /^[a-zA-Z0-9\s]+$/; // only chars, spaces and numbers 
     if (reg.test(req.query.name) && req.query.name.length <= 32){
         try {
-            await conn.query(`INSERT INTO recipe_books VALUES (null, ?, null, 1);`, [req.query.name]);
+            await conn.query(`INSERT INTO recipe_books VALUES (null, ?, ?, 1);`, [req.query.name, req.user.userId]);
         } catch (error) {
             console.log(error)
             res.sendStatus(500);
@@ -74,15 +87,18 @@ router.post("/create", async (req, res) => {
  *     description: "Book has been deleted"
  *   "400":
  *      description: "BookId parameter is wrong"
+ *    "403":
+ *      description: "Not allowed"
  */
-router.delete("/delete", async (req, res) => {
+router.delete("/delete", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
     res.status(400).send("You need to specify a bookId to delete");
     return;
   }
   try {
-    // Need to check if the req is the owner
-    await conn.query(`DELETE FROM recipe_books WHERE id=?;`, [req.query.bookId]);
+    if ((await bookExist(req.query.bookId))) return res.sendStatus(200);
+    if (!(await isOwner(req.query.bookId, req.user.userId))) return res.sendStatus(403);
+    await conn.query(`DELETE FROM recipe_books WHERE id=? AND owner = ?;`, [req.query.bookId, req.user.userId]);
   } catch (error) {
     console.log(error)
     res.sendStatus(500);
@@ -114,18 +130,21 @@ router.delete("/delete", async (req, res) => {
  *      description: "One of the parameter is wrong"
  *   "500":
  *      description: "The book you want to share does not exists"
+ *   "403":
+ *      description: "Not allowed"
  */
-router.post("/share", async (req, res) => {
+router.post("/share", needAuth, async (req, res) => {
   if (req.query.bookId === undefined || req.query.userId === undefined){
     res.status(400).send("Specify a bookId and the user you want to share the book with");
     return;
   }
   try {
-    // Need to check if the req is the owner
     if (await bookExist(req.query.bookId)){
       res.status(500).send("Can't share given book: no matching id");
       return;
-    }
+    }    
+    if (! (await isOwner(req.query.bookId, req.user.userId))) return res.sendStatus(403); // Status 500 if requester is not the owner
+    // TODO: Check for duplicates
     await conn.query(`INSERT INTO recipe_book_access VALUES (?,?);`, [req.query.bookId, req.query.userId]);
   } catch (error) {
     console.log(error)
@@ -149,12 +168,14 @@ router.post("/share", async (req, res) => {
  * responses:
  *   "200":
  *     description: "Book has been created"
+ *   "403":
+ *     description: "Problem with the read permission"
  *   "400":
  *     description: "Something wrong with bookId"
  *   "500":
  *      description: "The book does not exists"
  */
-router.get("/recipes", async (req, res) => {
+router.get("/recipes", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
       res.status(400).send("Need to specify a bookId");
       return;
@@ -164,6 +185,7 @@ router.get("/recipes", async (req, res) => {
         res.status(500).send("Can't get recipes from given book: no matching id");
         return;
       }
+      if (!(await userHasReadAccess(req.query.bookId, req.user.userId))) return res.sendStatus(403);
       const recipesRaw = await conn.query(`SELECT recipeId FROM recipe_book_links WHERE bookId = ?;`, [req.query.bookId]);
       let recipes = Array.from(recipesRaw);
       let ids = [];
@@ -195,10 +217,12 @@ router.get("/recipes", async (req, res) => {
  * responses:
  *   "200":
  *     description: "The general informations"
+ *   "403":
+ *     description: "User doesn't have access to the book"
  *   "400":
  *      description: "BookId parameter is wrong"
  */
-router.get("/general_information", async (req, res) => {
+router.get("/general_information", needAuth, async (req, res) => {
   if (req.query.bookId === undefined){
     res.status(400).send("You need to specify a bookId to search for");
     return;
@@ -209,7 +233,8 @@ router.get("/general_information", async (req, res) => {
         if (bookInfo.length <= 0){
             res.sendStatus(204); // No book found with the given id
             return;
-        // add an else if to know if we are authorized to read it
+        } else if (! (await userHasReadAccess(req.query.bookId, req.user.userId))){
+          return res.sendStatus(403);
         } else {
             let countRes =  await conn.query(`SELECT COUNT(bookId) FROM recipe_book_links WHERE bookId=?;`, [req.query.bookId]);
             let count = Number(countRes[0]['COUNT(bookId)']);
@@ -290,8 +315,10 @@ router.get("/id", async (req, res) => { //TODO: check for multiple names
  *     description: "The book does not exists"
  *   "400":
  *      description: "At least one parameter is wrong"
+ *   "403":
+ *      description: "Not allowed"
  */
-router.post("/recipe/add", async (req, res) => {
+router.post("/recipe/add", needAuth, async (req, res) => {
   if (req.query.bookId === undefined || req.query.recipeId === undefined){
     res.status(400).send("You need to specify a bookId and a recipeId");
     return;
@@ -303,7 +330,7 @@ router.post("/recipe/add", async (req, res) => {
       res.status(500).send("Can't add recipe to given book: no matching id");
       return;
     }
-    // need to check if requester is the owner of the book
+    if (!(await isOwner(req.query.bookId, req.user.userId))) return res.sendStatus(403);
     await conn.query(`INSERT INTO recipe_book_links VALUES (?, ?, ?);`, [req.query.bookId, req.query.recipeId, formattedDate]);
   } catch (error) {
     console.log(error);
@@ -334,14 +361,16 @@ router.post("/recipe/add", async (req, res) => {
  *     description: "Book has been deleted"
  *   "400":
  *      description: "At least one parameter is wrong"
+ *    "403":
+ *      description: "Not allowed"
  */
-router.delete("/share", async (req, res) => {
+router.delete("/share", needAuth, async (req, res) => {
   if (req.query.bookId === undefined || req.query.userId === undefined){
     res.status(400).send("You need to specify a bookId and a userId");
     return;
   }
   try {
-    // need to check if requester is the owner of the book
+    if (!(await isOwner(req.query.bookId, req.user.userId))) return res.sendStatus(403);
     await conn.query(`DELETE FROM recipe_book_access WHERE bookId = ? AND userId = ?;`, [req.query.bookId, req.query.userId]);
   } catch (error) {
     console.log(error);
